Making an SVG help file so I can use these functions in multiple benches
