+++ pages/benches/bench502.vue
@@ -1,7 +1,8 @@
 <script setup>
   import * as THREE from "three";
   import { OrbitControls } from "three/examples/jsm/controls/OrbitControls.js";
-  import { SVGLoader } from "three/examples/jsm/loaders/SVGLoader.js";
+  //   import { SVGLoader } from "three/examples/jsm/loaders/SVGLoader.js";
+  // import renderSVG_three
 
   definePageMeta({
     featured: false,
@@ -17,66 +18,6 @@
     // Load the SVG data from the file: @/data/usa-oh.svg
     const ohioSVG = await fetch("/assets/usa-oh.svg").then((res) => res.text());
     // console.log("SVG", ohioSVG);
-
-    const fillMaterial = new THREE.MeshBasicMaterial({ color: "#F3FBFB" });
-    const stokeMaterial = new THREE.LineBasicMaterial({
-      color: "#00A5E6"
-    });
-
-    const renderSVG = (extrusion, svg) => {
-      const loader = new SVGLoader();
-      const svgData = loader.parse(svg);
-      const svgGroup = new THREE.Group();
-      const updateMap = [];
-
-      svgGroup.scale.y *= -1;
-      svgData.paths.forEach((path) => {
-        const shapes = SVGLoader.createShapes(path);
-
-        shapes.forEach((shape) => {
-          const meshGeometry = new THREE.ExtrudeGeometry(shape, {
-            depth: extrusion,
-            bevelEnabled: false
-          });
-          const linesGeometry = new THREE.EdgesGeometry(meshGeometry);
-          const mesh = new THREE.Mesh(meshGeometry, fillMaterial);
-          const lines = new THREE.LineSegments(linesGeometry, stokeMaterial);
-
-          updateMap.push({ shape, mesh, lines });
-          svgGroup.add(mesh, lines);
-        });
-      });
-
-      const box = new THREE.Box3().setFromObject(svgGroup);
-      const size = box.getSize(new THREE.Vector3());
-      const yOffset = size.y / -2;
-      const xOffset = size.x / -2;
-
-      // Offset all of group's elements, to center them
-      svgGroup.children.forEach((item) => {
-        item.position.x = xOffset;
-        item.position.y = yOffset;
-      });
-      svgGroup.rotateX(-Math.PI / 2);
-
-      return {
-        object: svgGroup,
-        update(extrusion) {
-          updateMap.forEach((updateDetails) => {
-            const meshGeometry = new THREE.ExtrudeBufferGeometry(updateDetails.shape, {
-              depth: extrusion,
-              bevelEnabled: false
-            });
-            const linesGeometry = new THREE.EdgesGeometry(meshGeometry);
-
-            updateDetails.mesh.geometry.dispose();
-            updateDetails.lines.geometry.dispose();
-            updateDetails.mesh.geometry = meshGeometry;
-            updateDetails.lines.geometry = linesGeometry;
-          });
-        }
-      };
-    };
 
     // const loader = new SVGLoader();
     // const svgData = loader.parse(ohioSVG);
@@ -117,7 +58,7 @@
     scene.add(gridHelper);
 
     const defaultExtrusion = 1;
-    const { object, update } = renderSVG(defaultExtrusion, ohioSVG);
+    const { object, update } = renderSVG_three(defaultExtrusion, ohioSVG);
     scene.add(object);
 
     // Create the renderer
